+++ GIAO_DIEN/backActionChild/PolyBackAction.cs
@@ -4,8 +4,8 @@ namespace GIAO_DIEN.backActionChild
     class PolyBackAction : BackAction
     {
         public int polyName { get; set; }
-        public double mouseX { get; set; }
-        public double mouseY { get; set; }
+        public int mouseX { get; set; }
+        public int mouseY { get; set; }
 
         public override string ToString()
         {
